+++ controllers/contactController.js
@@ -13,7 +13,6 @@ const getAllContacts = async (req, res, next) => {
     res
       .status(HTTP_CODES.INTERNAL_SERVER_ERROR)
       .json({ ok: false, msg: 'internal server error', error });
-    console.log(error);
   }
 };
