+++ routes/auth.js
@@ -1,7 +1,7 @@
 const express = require('express');
 const router = express.Router();
 const { getUserData } = require('../controllers/authController');
-const { verifyUser } = require('../middleware/authMiddleware');
+const { verifyUser } = require('../middlewares/authMiddleware');
 
 router.get('/me', verifyUser, getUserData);
